refactor: separate pages Auth and Main on to components
