+++ src/pages/auth/Auth.tsx
@@ -1,6 +1,5 @@
+import { AuthForm, AuthLeftBlock } from '@/shared/components';
 import styles from './Auth.module.scss';
-import { AuthForm } from './ui/auth-form/AuthForm';
-import { AuthLeftBlock } from './ui/auth-left-block/AuthLeftBlock';
 
 export function Auth() {
 	return (
