+++ src/shared/components/auth-form/AuthForm.module.scss
