+++ src/shared/components/auth-left-block/AuthLeftBlock.module.scss
