+++ src/shared/components/auth-left-block/AuthLeftBlock.tsx
