+++ src/shared/components/content-top-left/ContentTopLeft.module.scss
