+++ src/shared/components/content-top-right/ContentTopRight.module.scss
@@ -1,16 +1,10 @@
 .contentTopRight {
   border-radius: 40px;
   width: 587px;
-  // width: auto;
-  // height: auto;
-  // background: #282828;
-  // padding: 30px 24px;
   position: relative;
-  // width: 203px;
 
   .mainImage {
     border-radius: 40px;
-    // width: 100%;
     height: 100%;
     object-fit: cover;
   }
