+++ src/shared/components/content-top/ContentTop.module.scss
