+++ src/shared/components/index.ts
@@ -1,0 +1,8 @@
+export { AuthForm } from './auth-form/AuthForm';
+export { AuthLeftBlock } from './auth-left-block/AuthLeftBlock';
+export { ContentTop } from './content-top/ContentTop';
+export { ContentTopLeft } from './content-top-left/ContentTopLeft';
+export { ContentTopRight } from './content-top-right/ContentTopRight';
+export { MainHeader } from './main-header/MainHeader';
+export { MainLogo } from './main-logo/MainLogo';
+export { MainMenu } from './main-menu/MainMenu';
