+++ src/shared/components/main-header/MainHeader.module.scss
@@ -17,7 +17,7 @@
     left: 20px;
     width: 24px;
     height: 24px;
-    background: transparent url(../../../../shared/assets/main/Search_24.svg) center center no-repeat;
+    background: transparent url(../../assets/main/Search_24.svg) center center no-repeat;
   }
   .headerSearch {
     padding-left: 58px;
@@ -42,7 +42,7 @@
   .navBell {
     width: 36px;
     height: 36px;
-    background: transparent url(../../../../shared/assets/main/Bell_24.svg) center center no-repeat;
+    background: transparent url(../../assets/main/Bell_24.svg) center center no-repeat;
   }
   .navAvatar {
     border-radius: 22px;
@@ -66,7 +66,7 @@
     .avatarArrow {
       width: 16px;
       height: 14px;
-      background: transparent url(../../../../shared/assets/main/Arrow_16.svg) center center no-repeat;
+      background: transparent url(../../assets/main/Arrow_16.svg) center center no-repeat;
     }
     .avatarFoto {
       position: absolute;
@@ -74,7 +74,7 @@
       left: 0px;
       width: 44px;
       height: 44px;
-      background: transparent url(../../../../shared/assets/main/Image-60.png) center center no-repeat;
+      background: transparent url(../../assets/main/Image-60.png) center center no-repeat;
     }
   }
 }
