+++ src/shared/components/main-logo/MainLogo.module.scss
