+++ src/shared/components/main-logo/MainLogo.tsx
