+++ src/shared/components/main-menu/MainMenu.module.scss
