+++ src/shared/components/main-menu/MainMenu.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 import navImgSrc from '@/shared/assets/main/Frame 30.png';
 import styles from './MainMenu.module.scss';
 import { MAIN_MENU_ITEMS } from '@/constants/constants';
-import { NavItem } from '../nav-item/NavItem';
+import { NavItem } from '@/shared/ui';
 
 export function MainMenu() {
   return (
