+++ src/shared/ui/divider/Divider.module.scss
