+++ src/shared/ui/index.ts
@@ -1,0 +1,4 @@
+export { Divider } from './divider/Divider';
+export { NavItem } from './nav-item/NavItem';
+export { Tag } from './tag/Tag';
+export { TopImage } from './top-image/TopImage';
