+++ src/shared/ui/nav-item/NavItem.module.scss
