+++ src/shared/ui/nav-item/NavItem.tsx
