+++ src/shared/ui/top-image/TopImage.module.scss
@@ -27,6 +27,6 @@
     right: 20px;
     width: 45px;
     height: 45px;
-    background: transparent url(../../../../shared/assets/main/arrowTop.svg) center center no-repeat;
+    background: transparent url(../../assets/main/arrowTop.svg) center center no-repeat;
   }
 }
