+++ src/shared/ui/top-image/TopImage.tsx
